Create a pull request template for Alcatraz
Sometimes reviewing PRs is hard when we don't fully understand the need or intent of a PR--basically, we're flying blind. The PR template is an effort to address this so that PRs can not only be understood better from the start, but will help us thoroughly test PRs, and hopefully help contributors think about more than Alcatraz itself (i.e. think about the primary use case, which is child theming).